Initial pass at modelling array job/tasks
There are now three types of jobs: `JOB`, `ARRAY_JOB` and `ARRAY_TASK`.

A `JOB` represents a non-array job, its submission script is submitted
to a single node allocated to it.

An `ARRAY_JOB` represents exactly that, an array job.  When created, it
will have a number of `ARRAY_TASK`s created for it.  The array task is
the job which is submitted to the daemon to be ran.  The state of the
array task can be tracked separately from the array job.

The scheduler does not concern itself with `ARRAY_TASK`, instead it
allocates resources and schedules `JOB`s and `ARRAY_JOB`s.  When an
`ARRAY_JOB` is ran its `ARRAY_TASK`s will be submitted to the allocated
nodes.

Limitations of array tasks:

Currently, only a single node allocated to an array job is used to run
the array tasks, and the tasks are not ran in parallel.  To fix these,
we need some way of tracking how many tasks can be ran on per-node and
which node has been allocated to which task.  Our allocation and job
models do not currently support.