Attempt to set the Priority reason when appropriate
As this is a FIFO queue the jobs will be ran in order. Thus if
the previous job has any reason it's not running, then the next
job's reason is Priority

There are likely a couple of edge cases where this fails, particularly
after a previous job has finished but the next hasn't started. But
these cases are brief and shouldn't cause an issue