Jobs have job steps
A job step represents a single parallel step for a job.  It consists of
an executable and arguments to execute.  The step is ran in parallel
over all nodes that have been allocated to the step's job.

The last point highlighted that the method we were using to initialize
and run batch jobs was not quite right.  Previously, we were
initializing the environment on only the first allocated node and
running the batch script on only that node.  This meant that any job
steps ran on other nodes would not have a correctly initialized
environment to run in.

Whilst looking at this I've also discovered that our model for array
jobs/tasks is not correct/consistent with slurm.

In slurm, when an array job is created with, say, 4 tasks to run across,
say, 2 nodes, *each* task is allocated 2 nodes.  If the nodes were
allocated exclusively and each task was to run at the same time, 8 nodes
would be allocated to the array job.

In flight scheduler by contrast, *each* task would be allocated a
*single* node, resulting in at most 4 nodes being allocated to the job.

This commit does not address that issue, but does lay the ground work by
adjusting `Submission::ArrayTask` to initialize all nodes that the array
task is allocated to.