Add JobRegistry
It maintains a registry of all jobs regardless of job type.  It also
provides a means to retrieve job tasks for an array job.

Currently, it is required to manually remove the jobs when they are no
longer wanted.  This is done well for `JOB` and `JOB_TASK`s but not for
`ARRAY_JOB`s.  This does not currently affect the scheduling, but it
will need to be addressed prior to job persistence being implemented.

A list of running tasks for an array job used to be maintained by the
`TaskRegistry` this is now queried from the `JobRegistry`.

The `FifoScheduler` now no longer maintains its own queue of jobs.
Instead it filters the jobs present in the job registry.