Replace `TaskRegistry` with `TaskGenerator`
Much of the task registry functionality has been removed with the
removal of `Job#to_enum` or merged into the new `JobRegistry`.

The remaining functionality is mostly centered around generating the
next task to be scheduled.  This functionality has been moved to a new
`ArrayTaskGenerator` class.

`ArrayTaskGenerator` generates the array tasks and provides methods to
query the state of the current generation.  Unlike `TaskRegistry` it
does not automatically advance its state instead advancing the state is
left to the scheduler as it is the scheduler that best knows if an array
task has been allocated resources.

With that refactoring `TaskRegistry` no longer serves a purposes and has
been removed.