Fix job registry deletion issue
`EventProcessor` no longer deletes jobs from the registry or cleans up
after jobs.  Instead it is responsible for updating the state of a job
and any dependent jobs.

`JobRegistry` has gained a new method to garbage collect any
completed/cancelled/failed JOB jobs and ARRAY_JOB jobs.  When it detects
a job that can be garbage collected, it cleans up the job before doing
so.

A new timer task has been written to periodically instruct `JobRegistry`
to garbage collect itself.  This is only ran when the production
webserver is started; it is not ran during tests.

In order to get the timer task working with the same `JobRegistry`
instance that the sinatra app uses, I've monkey patched Falcon to use a
threaded container.  Falcon really has reached the end of its
usefulness, but it can't be removed until we replace `Async::Websocket`
with some other bi-directional socket library.