ArrayTaskGenerator supports usage across fibres
It turns out that `Enumerator` doesn't like being called from different
`Fibre`s and therefore `Thread`s.  It raises an exception when this is
done.

In order to deserialize jobs, we will need be able to load the job and
therefore initialize the task generator from one fibre (or thread) and
use it from another.

We could try to work around that by making sure that we load the jobs in
the same fibre as is used by Falcon.  However, having
`ArrayTaskGenerator` support calls across threads is safer, easier and
more general.

Advancing the task pointed to by `ArrayTaskGenerator` is not thread
safe.  Advancing the task still needs to be synchronized, but there is
no exception when called by different threads.