Schedulers consider each partition independently
When attempting to allocate resources to a job, we iterate over the
partitions in the order that they are defined on the config file.  We
consider only jobs for the current partition.  When a job on the current
partition cannot be scheduled we ignore any more jobs on that partition
and move onto the next partition.