Add support for `.local` config files
Similar to the behaviour defined at
https://github.com/bkeepers/dotenv#what-other-env-files-can-i-use but
only for non-production environments.

In production, load only `etc/flight-scheduler-controller.yaml`.

In non-production environments, load
`etc/flight-scheduler-controller.#{env}.yaml` and
`etc/flight-scheduler-controller.#{env}.local.yaml`.  The values in
`*.local.yaml` taking precedence over the other config file.

This allows generic `development.yaml` and `test.yaml` files to be
committed to the repo and for each dev to have their own `*.local.yaml`
files that are not checked into the repo.

A configuration file not existing is no longer an issue.