Fix deallocation race condition
Introduced in 494a280, cleaning up completed ARRAY_JOBS didn't check
that its ARRAY_TASKs had been deallocated.  This resulted in the job
being removed whilst the allocation continued to live on.

This commit fixes the issue by checking ARRAY_TASKs.

We should also consider how to deal with de-allocations when the job is
not known.